Relocate `oneOf` from `helpers` to `src/` so that it can be imported by other libraries
