+++ src/index.js
@@ -2,5 +2,4 @@ export { default as compose } from './compose'
 export { default as createSpec } from './createSpec'
 export { default as faker } from './faker'
 export { default as derived } from './derived'
-
-export { default as oneOf } from './helpers/oneOf'
+export { default as oneOf } from './oneOf'
